POParser: support override of strings with custom po files
Default is that strings will not be overwritten so that custom po files
must be loaded first. Had to remove the previous behaviour of
substituting English where the translation file's English string
differs.
Added prints to warn of this and otherwise where strings differ so that
devs can correct if necessary.
A translation file will now overwrite a string entry that has only a
(matching) en_gb string (was previously ignored).